Move holiday CTA computation out of HolidayPeriodsContext
